modified some files

+++ 0-bubble_sort.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include <stdlib.h>
 #include "sort.h"
 
 /**
@@ -9,7 +8,7 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-	size_t i, j;
+	unsigned int i, j;
 	int tmp;
 
 	if (size < 2)
